requirements.txt
annotated-types==0.6.0
anyio==4.2.0
certifi==2024.8.30
click==8.1.7
distro==1.9.0
fastapi==0.112.2
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.7
openai==1.37.0
pip==24.2
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==0.21.0
setuptools==75.1.0
sniffio==1.3.0
starlette==0.38.2
tqdm==4.66.5
typing_extensions==4.11.0
uvicorn==0.20.0
wheel==0.44.0
